Move CallViewModel into the xml package
