+++ stream-video-android-xml/src/main/kotlin/io/getstream/video/android/xml/viewmodel/CallViewModel.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.getstream.video.android.common.viewmodel
+package io.getstream.video.android.xml.viewmodel
 
 import androidx.lifecycle.ViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
