+++ stream-video-android-xml/src/main/kotlin/io/getstream/video/android/xml/viewmodel/CallViewModelFactory.kt
@@ -14,12 +14,19 @@
  * limitations under the License.
  */
 
-package io.getstream.video.android.common.viewmodel
+package io.getstream.video.android.xml.viewmodel
 
-public interface CallViewModelFactoryProvider {
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
 
-    /**
-     * Allows for a custom implementation of the [CallViewModelFactory].
-     */
-    public fun getCallViewModelFactory(): CallViewModelFactory
+public class CallViewModelFactory() : ViewModelProvider.Factory {
+
+    public fun create(): CallViewModel {
+        return create(CallViewModel::class.java)
+    }
+
+    @Suppress("UNCHECKED_CAST")
+    override fun <T : ViewModel> create(modelClass: Class<T>): T {
+        return CallViewModel() as T
+    }
 }
